Cleaned up env files
